+++ src/core/interfaces/base-components-props.ts
@@ -1,19 +1,6 @@
 interface BaseComponentProps {
-  width?: string;
-  height?: string;
-  fontSize?: string;
-  fs?: string;
-  cursor?: string;
-  borderRadius?: string;
-  br?: string;
-  fontWeight?: string;
-  fw?: string;
-  border?: string;
-  transition?: string;
-  hover?: BaseComponentProps;
-  backgroundColor?: string;
-  bgColor?: string;
-  color?: string;
+  style?: string;
+  hover?: string;
 }
 
 export default BaseComponentProps;