+++ src/utils/valid-style-props.ts
@@ -1,0 +1,15 @@
+type validStylePropsType = {
+  [key: string]: string;
+};
+
+const validStyleProps: validStylePropsType = {
+  w: "width",
+  fs: "fontSize",
+  h: "height",
+  bgcolor: "backgroundColor",
+  cur: "cursor",
+  color: "color",
+  t: "transition",
+};
+
+export default validStyleProps;