+++ CHANGELOG.md
@@ -4,7 +4,19 @@ All notable changes to Skypiea UI will be documented in this file.
 
 ## [Unreleased]
 
-## [0.1.1] - 2022-04-29
+## [0.1.2] - 2022-04-29
+
+### Added
+
+- Ability to add base style by using a string
+- Ability to add hover style by using a string
+
+### Changed
+
+- Fixed backgroundColor not applying on hover
+- Refactoring of styling of component. No longer needed to give a prop for each style param
+
+## [0.1.1] - 2022-04-28
 
 ### Added